+++ linuxdoom-1.10/d_net.c
@@ -649,7 +649,6 @@ void TryRunTics (void)
     int		realtics;
     int		availabletics;
     int		counts;
-    int		numplaying;
     
     // get real tics		
     entertic = I_GetTime ()/ticdup;
@@ -660,12 +659,10 @@ void TryRunTics (void)
     NetUpdate ();
 	
     lowtic = MAXINT;
-    numplaying = 0;
     for (i=0 ; i<doomcom->numnodes ; i++)
     {
 	if (nodeingame[i])
 	{
-	    numplaying++;
 	    if (nettics[i] < lowtic)
 		lowtic = nettics[i];
 	}