+++ linuxdoom-1.10/p_enemy.c
@@ -1874,7 +1874,6 @@ void A_BrainAwake (mobj_t* mo)
     numbraintargets = 0;
     braintargeton = 0;
 	
-    thinker = thinkercap.next;
     for (thinker = thinkercap.next ;
 	 thinker != &thinkercap ;
 	 thinker = thinker->next)