+++ linuxdoom-1.10/r_data.c
@@ -250,8 +250,6 @@ void R_GenerateComposite (int texnum)
     colofs = texturecolumnofs[texnum];
     
     // Composite the columns together.
-    patch = texture->patches;
-		
     for (i=0 , patch = texture->patches;
 	 i<texture->patchcount;
 	 i++, patch++)
@@ -385,8 +383,6 @@ void R_GenerateLookup (int texnum)
     // "column without a patch" return below.
     byte		patchcount[texture->width];
     memset (patchcount, 0, texture->width);
-    patch = texture->patches;
-		
     for (i=0 , patch = texture->patches;
 	 i<texture->patchcount;
 	 i++, patch++)
@@ -492,7 +488,6 @@ void R_InitTextures (void)
     
     int*		patchlookup;
     
-    int			totalwidth;
     int			nummappatches;
     int			offset;
     int			maxoff;
@@ -567,8 +562,6 @@ void R_InitTextures (void)
     texturewidthmask = Z_Malloc (numtextures*sizeof(*texturewidthmask), PU_STATIC, 0);
     textureheight = Z_Malloc (numtextures*sizeof(*textureheight), PU_STATIC, 0);
 
-    totalwidth = 0;
-    
     //	Really complex printing shit...
     temp1 = W_GetNumForName ("S_START");  // P_???????
     temp2 = W_GetNumForName ("S_END") - 1;
@@ -650,8 +643,6 @@ void R_InitTextures (void)
 
 	texturewidthmask[i] = j-1;
 	textureheight[i] = texture->height<<FRACBITS;
-		
-	totalwidth += texture->width;
     }
 
     free (patchlookup);